Move http_ext module to the crate root
This was already its public path
