+++ apollo-router/src/services/mod.rs
@@ -14,8 +14,6 @@ use http::method::Method;
 use http::HeaderValue;
 use http::StatusCode;
 use http::Uri;
-use http_ext::IntoHeaderName;
-use http_ext::IntoHeaderValue;
 use multimap::MultiMap;
 use serde_json_bytes::ByteString;
 use static_assertions::assert_impl_all;
@@ -27,6 +25,9 @@ pub(crate) use self::supergraph_service::*;
 use crate::error::Error;
 use crate::graphql::Request;
 use crate::graphql::Response;
+use crate::http_ext;
+use crate::http_ext::IntoHeaderName;
+use crate::http_ext::IntoHeaderValue;
 use crate::json_ext::Object;
 use crate::json_ext::Path;
 use crate::json_ext::Value;
@@ -36,7 +37,6 @@ use crate::spec::Query;
 use crate::Context;
 
 mod execution_service;
-pub mod http_ext;
 pub(crate) mod layers;
 pub(crate) mod new_service;
 pub(crate) mod subgraph_service;
